feat: feedback

+++ src/cron/cleanZombieOpportunities.ts
@@ -11,7 +11,7 @@ export const cleanZombieOpportunities: Cron = {
   handler: async (con) => {
     const timeThreshold = subDays(new Date(), 2);
 
-    con.transaction(async (entityManager) => {
+    await con.transaction(async (entityManager) => {
       const query = entityManager
         .getRepository(Opportunity)
         .createQueryBuilder()
@@ -29,19 +29,29 @@ export const cleanZombieOpportunities: Cron = {
 
       const { affected, raw } = await query.execute();
 
-      const { affected: claimables } = await entityManager
-        .getRepository(ClaimableItem)
-        .createQueryBuilder()
-        .delete()
-        .where(`flags->>'opportunityId' IN (:...ids)`, {
-          ids: raw
-            .filter((item: { id: string }) => item)
-            .map((item: { id: string }) => item.id),
-        })
-        .execute();
+      const ids = raw
+        .filter((item: { id: string }) => item)
+        .map((item: { id: string }) => item.id);
+
+      let claimables = 0;
+
+      if (ids.length) {
+        const { affected } = await entityManager
+          .getRepository(ClaimableItem)
+          .createQueryBuilder()
+          .delete()
+          .where(`flags->>'opportunityId' IN (:...ids)`, {
+            ids,
+          })
+          .execute();
+
+        if (affected) {
+          claimables += affected;
+        }
+      }
 
       logger.info(
-        { count: affected, claimables },
+        { count: affected || 0, claimables },
         'zombies opportunities cleaned! 🧟',
       );
     });
